+++ src/canvas/canvas_events.ts
@@ -75,42 +75,42 @@ export class Canvas extends SelectableCanvas {
    * @type Number
    * @private
    */
-  mainTouchId: null | number;
+  declare mainTouchId: null | number;
 
   /**
    * When the option is enabled, PointerEvent is used instead of TPointerEvent.
    * @type Boolean
    * @default
    */
-  enablePointerEvents: boolean;
+  declare enablePointerEvents: boolean;
 
   /**
    * an internal flag that is used to remember if we already bound the events
    * @type Boolean
    * @private
    */
-  private eventsBound: boolean;
+  private declare eventsBound: boolean;
 
   /**
    * Holds a reference to a setTimeout timer for event synchronization
    * @type number
    * @private
    */
-  private _willAddMouseDown: number;
+  private declare _willAddMouseDown: number;
 
   /**
    * Holds a reference to an object on the canvas that is receiving the drag over event.
    * @type FabricObject
    * @private
    */
-  private _draggedoverTarget?: FabricObject;
+  private declare _draggedoverTarget?: FabricObject;
 
   /**
    * Holds a reference to an object on the canvas from where the drag operation started
    * @type FabricObject
    * @private
    */
-  private _dragSource?: FabricObject;
+  private declare _dragSource?: FabricObject;
 
   /**
    * Holds a reference to an object on the canvas that is the current drop target
@@ -119,11 +119,11 @@ export class Canvas extends SelectableCanvas {
    * @type FabricObject
    * @private
    */
-  private _dropTarget: FabricObject<ObjectEvents> | undefined;
+  declare private _dropTarget: FabricObject<ObjectEvents> | undefined;
 
-  currentTarget?: FabricObject;
+  declare currentTarget?: FabricObject;
 
-  currentSubTargets?: FabricObject[];
+  declare currentSubTargets?: FabricObject[];
 
   /**
    * Holds a reference to a pointer during mousedown to compare on mouse up and determine
@@ -131,7 +131,7 @@ export class Canvas extends SelectableCanvas {
    * @type FabricObject
    * @private
    */
-  _previousPointer: Point;
+  declare _previousPointer: Point;
 
   private _isClick: boolean;