+++ src/canvas/static_canvas.class.ts
@@ -87,7 +87,7 @@ export class StaticCanvas<
    * @type {(String|TFiller)}
    * @default
    */
-  backgroundColor: TFiller | string;
+  declare backgroundColor: TFiller | string;
 
   /**
    * Background image of canvas instance.
@@ -97,7 +97,7 @@ export class StaticCanvas<
    * @type FabricObject
    * @default
    */
-  backgroundImage: FabricObject | null;
+  declare backgroundImage: FabricObject | null;
 
   /**
    * Overlay color of canvas instance.
@@ -105,7 +105,7 @@ export class StaticCanvas<
    * @type {(String|TFiller)}
    * @default
    */
-  overlayColor: TFiller | string;
+  declare overlayColor: TFiller | string;
 
   /**
    * Overlay image of canvas instance.
@@ -115,7 +115,7 @@ export class StaticCanvas<
    * @type FabricObject
    * @default
    */
-  overlayImage: FabricObject | null;
+  declare overlayImage: FabricObject | null;
 
   /**
    * Indicates whether toObject/toDatalessObject should include default values
@@ -123,7 +123,7 @@ export class StaticCanvas<
    * @type Boolean
    * @default
    */
-  includeDefaultValues: boolean;
+  declare includeDefaultValues: boolean;
 
   /**
    * Indicates whether {@link add}, {@link insertAt} and {@link remove},
@@ -135,28 +135,28 @@ export class StaticCanvas<
    * @type Boolean
    * @default
    */
-  renderOnAddRemove: boolean;
+  declare renderOnAddRemove: boolean;
 
   /**
    * Indicates whether object controls (borders/controls) are rendered above overlay image
    * @type Boolean
    * @default
    */
-  controlsAboveOverlay: boolean;
+  declare controlsAboveOverlay: boolean;
 
   /**
    * Indicates whether the browser can be scrolled when using a touchscreen and dragging on the canvas
    * @type Boolean
    * @default
    */
-  allowTouchScrolling: boolean;
+  declare allowTouchScrolling: boolean;
 
   /**
    * Indicates whether this canvas will use image smoothing, this is on by default in browsers
    * @type Boolean
    * @default
    */
-  imageSmoothingEnabled: boolean;
+  declare imageSmoothingEnabled: boolean;
 
   /**
    * The transformation (a Canvas 2D API transform matrix) which focuses the viewport
@@ -167,7 +167,7 @@ export class StaticCanvas<
    * canvas.viewportTransform = [0.7, 0, 0, 0.7, 50, 50];
    * @default
    */
-  viewportTransform: TMat2D;
+  declare viewportTransform: TMat2D;
 
   /**
    * if set to false background image is not affected by viewport transform
@@ -176,7 +176,7 @@ export class StaticCanvas<
    * @todo we should really find a different way to do this
    * @default
    */
-  backgroundVpt: boolean;
+  declare backgroundVpt: boolean;
 
   /**
    * if set to false overlya image is not affected by viewport transform
@@ -185,14 +185,14 @@ export class StaticCanvas<
    * @todo we should really find a different way to do this
    * @default
    */
-  overlayVpt: boolean;
+  declare overlayVpt: boolean;
 
   /**
    * When true, canvas is scaled by devicePixelRatio for better rendering on retina screens
    * @type Boolean
    * @default
    */
-  enableRetinaScaling: boolean;
+  declare enableRetinaScaling: boolean;
 
   /**
    * Describe canvas element extension over design
@@ -202,7 +202,7 @@ export class StaticCanvas<
    * of canvas element in plain untrasformed coordinates
    * The coordinates get updated with @method calcViewportBoundaries.
    */
-  vptCoords: TCornerPoint;
+  declare vptCoords: TCornerPoint;
 
   /**
    * Based on vptCoords and object.aCoords, skip rendering of objects that
@@ -213,7 +213,7 @@ export class StaticCanvas<
    * @type Boolean
    * @default
    */
-  skipOffscreen: boolean;
+  declare skipOffscreen: boolean;
 
   /**
    * a fabricObject that, without stroke define a clipping area with their shape. filled in black
@@ -222,64 +222,69 @@ export class StaticCanvas<
    * clipPath will clip away controls, if you do not want this to happen use controlsAboveOverlay = true
    * @type FabricObject
    */
-  clipPath: FabricObject;
+  declare clipPath: FabricObject;
 
   /**
    * A reference to the canvas actual HTMLCanvasElement.
    * Can be use to read the raw pixels, but never write or manipulate
    * @type HTMLCanvasElement
    */
-  lowerCanvasEl: HTMLCanvasElement;
+  declare lowerCanvasEl: HTMLCanvasElement;
 
-  contextContainer: CanvasRenderingContext2D;
+  declare contextContainer: CanvasRenderingContext2D;
 
   /**
    * Width in virtual/logical pixels of the canvas.
    * The canvas can be larger than width if retina scaling is active
    * @type number
    */
-  width: number;
+  declare width: number;
 
   /**
    * Height in virtual/logical pixels of the canvas.
    * The canvas can be taller than width if retina scaling is active
    * @type height
    */
-  height: number;
+  declare height: number;
 
   /**
    * If true the Canvas is in the process or has been disposed/destroyed.
    * No more rendering operation will be executed on this canvas.
    * @type boolean
    */
-  destroyed?: boolean;
+  declare destroyed?: boolean;
 
   /**
    * Started the process of disposing but not done yet.
    * WIll likely complete the render cycle already scheduled but stopping adding more.
    * @type boolean
    */
-  disposed?: boolean;
+  declare disposed?: boolean;
 
   /**
    * Keeps a copy of the canvas style before setting retina scaling and other potions
    * in order to return it to original state on dispose
    * @type string
    */
-  _originalCanvasStyle?: string;
+  declare _originalCanvasStyle?: string;
 
-  renderAndResetBound: () => void;
-  requestRenderAllBound: () => void;
+  declare renderAndResetBound: () => void;
+  declare requestRenderAllBound: () => void;
 
-  _offset: { left: number; top: number };
-  protected hasLostContext: boolean;
-  protected nextRenderHandle: number;
+  declare _offset: { left: number; top: number };
+  protected declare hasLostContext: boolean;
+  protected declare nextRenderHandle: number;
 
   // reference to
-  protected __cleanupTask?: {
+  protected declare __cleanupTask?: {
     (): void;
     kill: (reason?: any) => void;
   };
+
+  constructor(el: string | HTMLCanvasElement, options = {}) {
+    super();
+    this._init(el, options);
+  }
 
   add(...objects: FabricObject[]) {
     const size = super.add(...objects);
@@ -323,11 +328,6 @@ export class StaticCanvas<
 
   _onStackOrderChanged() {
     this.renderOnAddRemove && this.requestRenderAll();
-  }
-
-  constructor(el: string | HTMLCanvasElement, options = {}) {
-    super();
-    this._init(el, options);
   }
 
   /**
@@ -1148,7 +1148,7 @@ export class StaticCanvas<
    * @type Boolean
    * @default
    */
-  svgViewportTransformation: boolean;
+  declare svgViewportTransformation: boolean;
 
   /**
    * Returns SVG representation of canvas