+++ src/shapes/itext.class.ts
@@ -68,53 +68,53 @@ export class IText extends ITextClickBehaviorMixin<ITextEvents> {
    * @type Number
    * @default
    */
-  selectionStart: number;
+  declare selectionStart: number;
 
   /**
    * Index where text selection ends
    * @type Number
    * @default
    */
-  selectionEnd: number;
+  declare selectionEnd: number;
 
-  compositionStart: number;
+  declare compositionStart: number;
 
-  compositionEnd: number;
+  declare compositionEnd: number;
 
   /**
    * Color of text selection
    * @type String
    * @default
    */
-  selectionColor: string;
+  declare selectionColor: string;
 
   /**
    * Indicates whether text is in editing mode
    * @type Boolean
    * @default
    */
-  isEditing: boolean;
+  declare isEditing: boolean;
 
   /**
    * Indicates whether a text can be edited
    * @type Boolean
    * @default
    */
-  editable: boolean;
+  declare editable: boolean;
 
   /**
    * Border color of text object while it's in editing mode
    * @type String
    * @default
    */
-  editingBorderColor: string;
+  declare editingBorderColor: string;
 
   /**
    * Width of cursor (in px)
    * @type Number
    * @default
    */
-  cursorWidth: number;
+  declare cursorWidth: number;
 
   /**
    * Color of text cursor color in editing mode.
@@ -124,28 +124,28 @@ export class IText extends ITextClickBehaviorMixin<ITextEvents> {
    * @type String
    * @default
    */
-  cursorColor: string;
+  declare cursorColor: string;
 
   /**
    * Delay between cursor blink (in ms)
    * @type Number
    * @default
    */
-  cursorDelay: number;
+  declare cursorDelay: number;
 
   /**
    * Duration of cursor fade in (in ms)
    * @type Number
    * @default
    */
-  cursorDuration: number;
+  declare cursorDuration: number;
 
   /**
    * Indicates whether internal text char widths can be cached
    * @type Boolean
    * @default
    */
-  caching: boolean;
+  declare caching: boolean;
 
   /**
 
@@ -627,9 +627,9 @@ export const iTextDefaultValues: Partial<TClassProperties<IText>> = {
   cursorDuration: 600,
   caching: true,
   hiddenTextareaContainer: null,
-  _reSpace: /\s|\n/,
   _currentCursorOpacity: 1,
   _selectionDirection: null,
+  _reSpace: /\s|\n/,
   inCompositionMode: false,
   keysMap,
   keysMapRtl,